requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

einops
hydra-core
lightning
lion_pytorch
matplotlib
numpy
omegaconf
pandas
pytorch_lightning
Requests
scikit_learn
scipy
seaborn
torch==2.5.0+cu121
torch_ema
transformers
wandb
kagglehub
onnx
onnxruntime-gpu
backtesting
